move construction of akka stream from PeerManager to NeutrinoNode
